Move TransactionTagCounter into separate file
