+++ fdbserver/CMakeLists.txt
@@ -142,6 +142,7 @@ set(FDBSERVER_SRCS
   TesterInterface.actor.h
   TLogInterface.h
   TLogServer.actor.cpp
+  TransactionTagCounter.h
   TSSMappingUtil.actor.h
   TSSMappingUtil.actor.cpp
   VersionedBTree.actor.cpp
