+++ fdbserver/storageserver.actor.cpp
@@ -60,6 +60,7 @@
 #include "fdbserver/StorageMetrics.h"
 #include "fdbserver/ServerDBInfo.h"
 #include "fdbserver/TLogInterface.h"
+#include "fdbserver/TransactionTagCounter.h"
 #include "fdbserver/WaitFailure.h"
 #include "fdbserver/WorkerInterface.actor.h"
 #include "fdbrpc/sim_validation.h"
@@ -741,78 +742,6 @@ public:
 		maxQueryQueue = 0;
 		return val;
 	}
-
-	struct TransactionTagCounter {
-		struct TagInfo {
-			TransactionTag tag;
-			double rate;
-			double fractionalBusyness;
-
-			TagInfo(TransactionTag const& tag, double rate, double fractionalBusyness)
-			  : tag(tag), rate(rate), fractionalBusyness(fractionalBusyness) {}
-		};
-
-		TransactionTagMap<int64_t> intervalCounts;
-		int64_t intervalTotalSampledCount = 0;
-		TransactionTag busiestTag;
-		int64_t busiestTagCount = 0;
-		double intervalStart = 0;
-
-		Optional<TagInfo> previousBusiestTag;
-
-		UID thisServerID;
-
-		Reference<EventCacheHolder> busiestReadTagEventHolder;
-
-		TransactionTagCounter(UID thisServerID)
-		  : thisServerID(thisServerID),
-		    busiestReadTagEventHolder(makeReference<EventCacheHolder>(thisServerID.toString() + "/BusiestReadTag")) {}
-
-		int64_t costFunction(int64_t bytes) const { return bytes / SERVER_KNOBS->READ_COST_BYTE_FACTOR + 1; }
-
-		void addRequest(Optional<TagSet> const& tags, int64_t bytes) {
-			if (tags.present()) {
-				TEST(true); // Tracking tag on storage server
-				double cost = costFunction(bytes);
-				for (auto& tag : tags.get()) {
-					int64_t& count = intervalCounts[TransactionTag(tag, tags.get().getArena())];
-					count += cost;
-					if (count > busiestTagCount) {
-						busiestTagCount = count;
-						busiestTag = tag;
-					}
-				}
-
-				intervalTotalSampledCount += cost;
-			}
-		}
-
-		void startNewInterval() {
-			double elapsed = now() - intervalStart;
-			previousBusiestTag.reset();
-			if (intervalStart > 0 && CLIENT_KNOBS->READ_TAG_SAMPLE_RATE > 0 && elapsed > 0) {
-				double rate = busiestTagCount / CLIENT_KNOBS->READ_TAG_SAMPLE_RATE / elapsed;
-				if (rate > SERVER_KNOBS->MIN_TAG_READ_PAGES_RATE) {
-					previousBusiestTag = TagInfo(busiestTag, rate, (double)busiestTagCount / intervalTotalSampledCount);
-				}
-
-				TraceEvent("BusiestReadTag", thisServerID)
-				    .detail("Elapsed", elapsed)
-				    .detail("Tag", printable(busiestTag))
-				    .detail("TagCost", busiestTagCount)
-				    .detail("TotalSampledCost", intervalTotalSampledCount)
-				    .detail("Reported", previousBusiestTag.present())
-				    .trackLatest(busiestReadTagEventHolder->trackingKey);
-			}
-
-			intervalCounts.clear();
-			intervalTotalSampledCount = 0;
-			busiestTagCount = 0;
-			intervalStart = now();
-		}
-
-		Optional<TagInfo> getBusiestTag() const { return previousBusiestTag; }
-	};
 
 	TransactionTagCounter transactionTagCounter;
 
@@ -3645,9 +3574,9 @@ void getQueuingMetrics(StorageServer* self, StorageQueuingMetricsRequest const& 
 	reply.diskUsage = self->diskUsage;
 	reply.durableVersion = self->durableVersion.get();
 
-	Optional<StorageServer::TransactionTagCounter::TagInfo> busiestTag = self->transactionTagCounter.getBusiestTag();
-	reply.busiestTag = busiestTag.map<TransactionTag>(
-	    [](StorageServer::TransactionTagCounter::TagInfo tagInfo) { return tagInfo.tag; });
+	Optional<TransactionTagCounter::TagInfo> busiestTag = self->transactionTagCounter.getBusiestTag();
+	reply.busiestTag =
+	    busiestTag.map<TransactionTag>([](TransactionTagCounter::TagInfo tagInfo) { return tagInfo.tag; });
 	reply.busiestTagFractionalBusyness = busiestTag.present() ? busiestTag.get().fractionalBusyness : 0.0;
 	reply.busiestTagRate = busiestTag.present() ? busiestTag.get().rate : 0.0;
